+++ CalendarApi.sln
@@ -1,6 +1,6 @@
 ﻿
 Microsoft Visual Studio Solution File, Format Version 12.00
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Api", "Api\Api.csproj", "{EA7E910D-D5FE-4EDD-ADB4-F4C567C74020}"
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "CalendarApi", "CalendarApi\CalendarApi.csproj", "{EA7E910D-D5FE-4EDD-ADB4-F4C567C74020}"
 EndProject
 Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "Solution Items", "Solution Items", "{6DD0F012-026F-4869-829D-CA527B7A3892}"
 	ProjectSection(SolutionItems) = preProject
