+++ CalendarApi/Models/Animal.cs
