+++ CalendarApi/Models/Appointment.cs
